feature[product]: create product

+++ database/migrations/2024_12_21_222519_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('category');
-            $table->decimal('price', 10, 2);
+            $table->int('price');
             $table->string('sku');
             $table->timestamps();
         });
